Refactor code in `tune_katago.py`
This commit modifies the import statement to import the `bernoulli` function from the `scipy.stats` module. It also adds the `math` module for mathematical calculations. The `translate_parameters` and `translate_solutions` functions are updated to use the new ordering and eliminate unnecessary parameters. The `get_katago_parameters` function is modified to use a different value for the "maxVisits" parameter. The `plot_pairwiseclop` function is added to plot the evolution of optimums. The `run_pairwiseclop` function is updated to call the `plot_pairwiseclop` function. New functions `elo` and `elo_range` are added to calculate expected ELO and ELO standard deviation. The `tune` function is modified to add the option of plotting graphs and print the expected ELO and ELO range. The `simulation` and `plotting` variables are set to `True` to enable simulation and graph plotting.